FIX-#3252: refactor 'handle_as_index'
